+++ alikafkastreaming-20260202/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-06-15 Version: 1.0.1
+- Update API CreateComputeInstance: add request parameters ResourceGroupId.
+- Update API CreateComputeInstance: add request parameters ResourceType.
+
+
 2026-06-12 Version: 1.0.0
 - Generated python 2026-02-02 for AlikafkaStreaming.